// HealthCare/src/Interfaces/ClinicMaster/clinic-request.ts
export interface ClinicRequest {
  clinicId: number;
  clinicCode: string;
  clinicName: string;
  clinicIncharge: string;
  emailId: string;
  mobileNumber: string;
  alterContactNumber: string;
  clinicDoctorCode: string;
  centerCode: string;
  clinicAddress: string;
  rateType: string;
  clinicStatus: boolean;
  createdBy: string;
  updatedBy: string;
  partnerId: string;
}